requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pythorhead==0.20.0
PyYAML==6.0.1
requests==2.31.0
semver==3.0.2
urllib3==2.1.0
